backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.10.0
pydantic-settings>=2.1.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
PyPDF2==3.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
